core: Bound number of async imports by CPUs
I was doing a `--ex-unified-core` compose of FAW, and started hitting
the fd limit (1024), since we created hundreds of threads.  Since
it doesn't make sense to do that, let's bound the number of concurrent
imports by the number of CPUs.

This is another implementation of the "bounded async" pattern, like
https://github.com/ostreedev/ostree/commit/c18628ecb8b2ef591db0444fa57052cba60807a8

Closes: #1317
Approved by: jlebon